requirements.txt
asgiref==3.8.1
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.0.7
django-cors-headers==4.9.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
psycopg2-binary==2.9.11
pyserial==3.5
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.5.0
whitenoise==6.11.0
